requirements.txt
python-telegram-bot
openai-whisper
torch
